reorganized header components and abstracted apps header
