--- components/apps/header.module.scss
@@ -1,10 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: center;
-
-  h1 {
-    margin-left: 1rem;
-  }
-}
